requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.23
psycopg2-binary>=2.9.9  # Use psycopg2 for Railway PostgreSQL compatibility
aiosqlite>=0.19.0  # SQLite support for fallback
alembic>=1.12.1

# Redis (for caching/sessions)
redis>=5.0.1

# Telegram Bot
python-telegram-bot>=20.7

# HTTP client
httpx>=0.25.2
aiohttp>=3.9.1

# Image processing
pillow>=10.1.0
pillow-heif>=0.14.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# AI/ML
google-generativeai>=0.3.2

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.3
python-dateutil>=2.8.2

# Background tasks
APScheduler>=3.10.4

# Monitoring
prometheus-client>=0.19.0
